BackEnd/requirements.txt
acres==0.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
billiard==4.2.1
blis==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
cymem==2.0.11
dnspython==2.7.0
docx==0.2.4
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
etelemetry==0.3.1
fastapi==0.115.13
filelock==3.18.0
firebase-admin==6.9.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
indian-namematch==1.3.0
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
murmurhash==1.0.13
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pathlib==1.0.1
pdfminer.six==20250506
pillow==11.2.1
preshed==3.0.10
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.2
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdfium2==4.30.1
pyphonetics==0.5.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
traits==7.0.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.13
uvicorn==0.34.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zipfile36==0.1.3
